requirements.txt
SpeechRecognition
google-cloud-speech
pyautogui
dotenv
keyboard
cv2
opencv-python
mediapipe
pywin32
julep
pyyaml